requirements.txt
requests
datetime
python-dateutil
discord
python-dotenv
beautifulsoup4
mysql-connector-python